Update fileshred.py
Changed: It will now show the file sizes in megabytes in the list of files matched and on the file currently processing. This way, the use can estimate how long it might take to overwrite.

Changed: When the file is deleted, it will now say "deleted" instead of "destroyed". When it says "destroyed", the file is just destroyed but still exists in the FS (this is when --no-delete is used, and is useful for debugging purposes)